accufiz_interferometer now following CameraProxy service interface

+++ catkit2/services/accufiz_interferometer_sim/accufiz_interferometer_sim.py
@@ -9,15 +9,14 @@ import math
 from astropy.io import fits
 from glob import glob
 from catkit2.testbed.service import Service
-from catkit2.testbed.service_proxy import ServiceProxy
 from catkit2.utils import rotate_and_flip_image
 import tempfile
 import os
+import threading
+
 
 class sim_response:
     text = 'success'
-
-    import numpy as np
 
 def generate_circle_array(radius=1, h=256, w=256):
     # radius is the proportion of the circle (0.5 is half the size)
@@ -39,7 +38,6 @@ def generate_circle_array(radius=1, h=256, w=256):
     return circle_array.astype('uint8')
 
 
-@ServiceProxy.register_service_interface('accufiz_interferometer_sim')
 class AccufizInterferometerSim(Service):
     NUM_FRAMES_IN_BUFFER = 20
 
@@ -61,6 +59,8 @@ class AccufizInterferometerSim(Service):
         self.image_height = self.config.get('height', 1967)
         self.image_width = self.config.get('width', 1970)
         self.config_id = self.config.get('config_id', 'accufiz')
+        self.save_h5 = self.config.get('save_h5', True)
+        self.save_fits = self.config.get('save_fits', False)
 
         # Set the 4D timeout.
         self.html_prefix = f"http://{self.ip}/WebService4D/WebService4D.asmx"
@@ -72,10 +72,16 @@ class AccufizInterferometerSim(Service):
         # Create data streams.
         self.detector_masks = self.make_data_stream('detector_masks', 'uint8', [self.image_height, self.image_width], self.NUM_FRAMES_IN_BUFFER)
         self.images = self.make_data_stream('images', 'float32', [self.image_height, self.image_width], self.NUM_FRAMES_IN_BUFFER)
+        
+        self.is_acquiring = self.make_data_stream('is_acquiring', 'int8', [1], 20)
+        self.is_acquiring.submit_data(np.array([0], dtype='int8'))
+        self.should_be_acquiring = threading.Event()
+        self.should_be_acquiring.clear()
 
         self.make_command('take_measurement', self.take_measurement)
-    
-    
+        self.make_command('start_acquisition', self.start_acquisition)
+        self.make_command('end_acquisition', self.end_acquisition)
+
     instrument_lib = requests
 
     def set_mask(self):
@@ -94,12 +100,13 @@ class AccufizInterferometerSim(Service):
     def get(self, url, params=None, **kwargs):
         return sim_response()
 
+
     def post(self, url, data=None, json=None, **kwargs):
         time.sleep(self.post_save_sleep)
         return sim_response()
 
 
-    def take_measurement(self, num_frames=2):
+    def take_measurement(self, num_frames=1):
         # Send request to take data.
         resp = self.post(f"{self.html_prefix}/AverageMeasure", data={"count": int(num_frames)})
         if "success" not in resp.text:
@@ -142,18 +149,23 @@ class AccufizInterferometerSim(Service):
         img = np.array(h5py.File(local_file_path, 'r').get('measurement0').get('genraw').get('data')) * mask
 
         self.detector_masks.submit_data(mask.astype(np.uint8))
-        self.images.submit_data(img.astype(np.float32))
+        #self.images.submit_data(img.astype(np.float32))
 
-        image = self.convert_h5_to_fits(local_file_path, rotate=0, fliplr=True, mask=mask, img=img)
+        image = self.convert_h5_to_fits(local_file_path, rotate=0, fliplr=True, mask=mask, img=img, create_fits=self.save_fits)
         if temp_file:
-            os.remove(local_file_path)
-            os.remove(local_file_path.replace('.h5', '.fits'))
+            fitsfile = local_file_path.replace('.h5', '.fits')
+
+            if os.path.exists(local_file_path):
+                os.remove(local_file_path)
+            
+            if os.path.exists(fitsfile):
+                os.remove(fitsfile)
             self.log.info('cleaning up temporary simulated files')
         return image
 
 
     @staticmethod
-    def convert_h5_to_fits(filepath, rotate, fliplr, img, mask, wavelength=632.8):
+    def convert_h5_to_fits(filepath, rotate, fliplr, img, mask, wavelength=632.8, create_fits=False):
 
         filepath = filepath if filepath.endswith(".h5") else f"{filepath}.h5"
 
@@ -168,22 +180,50 @@ class AccufizInterferometerSim(Service):
         center = ndimage.measurements.center_of_mass(mask)
 
         image = np.clip(img, -10, +10)[np.int64(center[0]) - radiusmask:np.int64(center[0]) + radiusmask - 1,
-                np.int64(center[1]) - radiusmask: np.int64(center[1]) + radiusmask - 1]
+                        np.int64(center[1]) - radiusmask: np.int64(center[1]) + radiusmask - 1]
 
         # Apply the rotation and flips.
         image = rotate_and_flip_image(image, rotate, fliplr)
 
         # Convert waves to nanometers.
         image = image * wavelength
-
-        fits_hdu = fits.PrimaryHDU(image)
-        fits_hdu.writeto(fits_filepath, overwrite=True)
+        if create_fits:
+            fits_hdu = fits.PrimaryHDU(image)
+            fits_hdu.writeto(fits_filepath, overwrite=True)
         return image
 
     def main(self):
         while not self.should_shut_down:
-            self.sleep(0.1)
+            if self.should_be_acquiring.wait(0.05):
+                self.acquisition_loop()
 
+    def acquisition_loop(self):
+        try:
+            self.is_acquiring.submit_data(np.array([1], dtype='int8'))
+
+            while self.should_be_acquiring.is_set() and not self.should_shut_down:
+                img = self.take_measurement()
+
+                # Make sure the data stream has the right size and datatype.
+                has_correct_parameters = np.allclose(self.images.shape, img.shape)
+
+                if not has_correct_parameters:
+                    self.images.update_parameters('uint16', img.shape, 20)
+                self.log.info('requesting new frame')
+                frame = self.images.request_new_frame()
+
+                frame.data[:] = img
+                self.log.info(f'frame {frame.id} acquired')
+                self.images.submit_frame(frame.id)
+                time.sleep(0.05)
+        finally:
+            self.is_acquiring.submit_data(np.array([0], dtype='int8'))
+
+    def start_acquisition(self):
+        self.should_be_acquiring.set()
+
+    def end_acquisition(self):
+        self.should_be_acquiring.clear()
 
 if __name__ == '__main__':
     service = AccufizInterferometerSim()
